fix issues

+++ src/main/java/io/springworks/service/BookService.java
@@ -1,12 +1,15 @@
 package io.springworks.service;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import io.springworks.models.Book;
 import io.springworks.repos.BookRepo;
@@ -21,21 +24,21 @@ public class BookService {
 
 	private List<Book> books;
 
-	public List<Book> getBooks(){
-		books = new ArrayList<>();
-		bookRepo.findAll()
-		.forEach(books::add);
-		logger.info("Found {} books", books.size());
-		return books;
+	public long getBooksCount(){
+		long count = bookRepo.count();
+		logger.info("Found {} books", count);
+		return count;
+
 	}
 
-	public Book getbook(int id) {
-		Object book = bookRepo.findById(id);
-		if(book != null) {
+	public List<Book> getbook(int id) {
+		List<Book> books = new ArrayList<>();
+		Optional<Book> book = bookRepo.findById(id);
+		if(book.isPresent()) {
 			logger.info("Found book with ID: {}", id);
-			return (Book) book;
+			books.add(book.get());
 		}
-		return null;
+		return books;
 	}
 
 	public void addBook(Book book) {
@@ -43,8 +46,14 @@ public class BookService {
 		bookRepo.save(book);
 	}
 
-	public void deleteBook(Book book) {
-		logger.info("Deleting book by Author: {} Book name - {} from inventory", book.getAuthor(), book.getName());
-		bookRepo.delete(book);
+	public void deleteBook(int id) {
+		List<Book> books = getbook(id);
+		if(!CollectionUtils.isEmpty(books)){
+			Book book = books.get(0);
+			bookRepo.delete(book);
+			logger.info("Deleting book by Author: {} Book name - {} from inventory", book.getAuthor(), book.getName());
+			return;
+		}
+		logger.error("Could not find the book by ID: {}", id);
 	}
 }
